requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
gspread==6.0.0
google-auth==2.23.4
twilio==8.10.0
requests>=2.31.0
python-dotenv==1.0.0